src/backend/requirements.txt
blinker==1.8.2
click==8.1.7
eclipse-zenoh==1.0.3
flask==3.0.3
Flask-Cors==5.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
werkzeug==3.0.6
zipp==3.20.2
actionlib==1.13.2
angles==1.9.13
catkin==0.8.8
click==8.1.7
diagnostic-updater==1.10.2
eclipse-zenoh==0.11.0
Flask==2.2.5
Flask-Cors==5.0.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
mavros==1.4.0
message-filters==1.15.8
pkg-resources==0.0.0
rosbag==1.15.8
rosboost-cfg==1.15.6
rosclean==1.15.6
roscreate==1.15.6
rosgraph==1.15.8
roslaunch==1.15.8
roslib==1.15.6
roslz4==1.15.8
rosmake==1.15.6
rosmaster==1.15.8
rosmsg==1.15.8
rosnode==1.15.8
rosparam==1.15.8
rospy==1.15.8
rosservice==1.15.8
rostest==1.15.8
rostopic==1.15.8
rosunit==1.15.6
roswtf==1.15.8
sensor-msgs==1.13.0
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.8
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
